added some strings, changed some functions' return types to ZunResult, renamed FUN_0043a1ec to LoadReplayMenu, renamed FUN_004051b0 to ExecuteAnmIdx

+++ src/th06_102h.exe/MainMenu/OnUpdate.c
@@ -612,7 +612,7 @@ LAB_0043666d:
   case 0xd:
   case 0xe:
   case 0xf:
-    FUN_004381ec((int)param_1);
+    FUN_004381ec(param_1);
     if (uVar3 != 0) {
       return 0;
     }
